Добавит в `provider` проверку на доступность интернета

+++ src/api/provider.js
@@ -1,25 +1,54 @@
+const isOnline = () => {
+  return window.navigator.onLine;
+};
+
 export default class Provider {
   constructor(api) {
     this._api = api;
   }
 
   getMovies() {
-    return this._api.getMovies();
+    if (isOnline()) {
+      return this._api.getMovies();
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   getComments(movieId) {
-    return this._api.getComments(movieId);
+    if (isOnline()) {
+      return this._api.getComments(movieId);
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   createComment(filmId, data) {
-    return this._api.createComment(filmId, data);
+    if (isOnline()) {
+      return this._api.createComment(filmId, data);
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   updateMovie(id, data) {
-    return this._api.updateMovie(id, data);
+    if (isOnline()) {
+      return this._api.updateMovie(id, data);
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   deleteComment(id) {
-    return this._api.deleteComment(id);
+    if (isOnline()) {
+      return this._api.deleteComment(id);
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 }
